backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "AI Education Platform Backend - FastAPI + Kimi CLI Integration"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "websockets>=14.0",
    "kimi-cli",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    # Video processing dependencies
    "openai-whisper>=20240930",
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.8.0",
    "reportlab>=4.0.0",
    "pillow>=10.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["video_processor*"]

[tool.uv.sources]
kimi-cli = { path = "../kimi-cli", editable = true }
